pyproject.toml
[project]
name = "youposm"
version = "1.0.0"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "gspread>=5.11.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-secret-manager>=2.16.0",
    "google-auth>=2.23.0",
    "google-oauth2-tool>=0.0.3",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "app-local", "debug-diagnosis"]